Bug 12485 - Show OPAC search history when opacuserlogin is OFF
In the boostrap OPAC the link to search history is missing if
opacuserlogin is turned off. OPAC search history should be available
whether or not users can log in. This patch removes this dependency.

To test, apply the patch and test with various combinations of
opacuserlogin and EnableOpacSearchHistory:

- opacuserlogin and EnableOpacSearchHistory both enabled
- opacuserlogin enabled and EnableOpacSearchHistory disabled
- opacuserlogin disabled and EnableOpacSearchHistory enabled

Followed test plan. Works as expected.
Signed-off-by: Marc Véron <veron@veron.ch>

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Works as described, no problems found.
Passes QA script and tests.

Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>