Bug 12330: [QA Follow-up] Consistency between opac-search and staff client
In opac-search two regexes are added for the results of splitting the
pref AdvancedSearchTypes, allowing spaces before and after the names.
The same regexes should be added to catalogue/search.pl to have the
same behavior there.

Test plan:
Set AdvancedSearchTypes to "itemtypes  |  ccode |  loc " (no quotes).
Check OPAC and staff with and without this patch.

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>