Bug 12574 - BiblioDefaultView doesn't affect XSLT results view
The OPAC's XSLT results view does not respect the choice set in the
BiblioDefaultView system preference. This patch adds a check of the
preference to the OPAC's results view for MARC21.

To test you must have the OPAC configured to use the XSLT view for
search results. Perform any search and test the link to a title's detail
page. Switching the values of the BiblioDefaultView should be correctly
reflected in the search results links.

Signed-off-by: Nick Clemens <nick@quecheelibrary.org>
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Works as described for XSLT view in OPAC.

Notes:
- maybe we should change the pref text from 'simple form' to 'normal view'.
- the patch doesn't include changes to the title link generated from an 880
  field - so those links will still lead to the wrong view.

Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>