Bug 9013 - Cart in staff interface displays location pulled from Description (OPAC)
When the cart in the staff client pulls description data for
items.location and items.ccode it pulls the value designated for the
OPAC--probably a cut-and-paste error when the feature was added to the
staff client.

This patch removes the 'opac' parameter from GetKohaAuthorisedValues.

To test you must have at least one LOC authorized value configured to
have different descriptions for staff and OPAC. Locate or edit an item
with that shelving location and add that record to the cart in the staff
client.

When you view the cart, the shelving location description in the "Items"
column should show the description for the staff client, not the OPAC.

Signed-off-by: Nick Clemens <nick@debian>
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Works as described, no problems found.

Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>