Bug 12589 - Manage label batch view should show item type description instead of code
When viewing the list of items in an existing label batch the item type
code is shown. This patch modifies the template to show the item type
description instead.

To test, go to Tools -> Labels -> Manage batches. If necessary, create a
batch with multiple items of different item types. Edit the batch and
confirm that the table of items shows item type description instead of
code.

Signed-off-by: Robert Higgins <robert.higgins@catalyst.net.nz>

Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>