Bug 12614: Remove CGI::scrolling_list from auth_subfields_structure.pl
This patch removes all instances on this file. Even removes some commented out instances.

Also re-enable editor feature to show/change kohafield value, can't find when or why it was removed.
Up to QA or RM for consideration, simply to remove.
But is strange to have a non editable field on auth frameworks.

To test:
1. Apply the patch
2. Go to Administration > Authority types
3. Clic MARC structure of any auth fw
4. Clic subfield on any tag, e.g. 031
5. Clic 'Edit subfields' button

Now the tests proper

6. 'Help input' box (editor bottom):
Eight(8) scrolling_list removed, they use to
feed the four pulldowns, so check each one
'Koha field' is new (for me at last)
'Authorised value',
'Thesaurus' (is this valid/useful here?)
'Plugin'
There is a difference between existing subtags and a new subtag,
the 'id' of the pulldowns (so the 8 scrolling list removed)
Check assign/remove/save/load values, search for regressions
If you edit subtag 942, it has a value on kohafield (default fw)

I think that this box could be prettified with fixed width for
pulldowns.

7. No more functional tests, there are 4 scrolling_list removed
that are commented (removed also commented code on tt file),
and two instances that are similar to previously removed
instances, evidently with no effect.

Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Not sure about the new Koha field entry... All works as expected.

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
All works as expected.
About Koha field: I think it makes sense to have it in the form,
so you can configure the fields for new authority types.
In my database I have mappings on authtypecode and authid.

Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>