Bug 12597 - Give better visual indication of currently-selected language in the staff client
There is currently no visual feedback to the user reflecting which of
multiple installed translations is currently selected. I think some
necessary styling didn't make it into the conversion to Bootstrap menus.
This patch attempts to improve things.

Most of the changes are to bring back the previous style of highlighting
the currently selected language. A new change is the addition of the
sub-language code to the display of the menu label if one of that menu
item's sub-languages is selected.

This patch also removes some obsolete CSS which referred to the old menu
system.

To test, apply the patch and install or update multiple languages,
including at least two sub-languages. In my test I used en, en-GB and
de-DE. Clear your browser cache if necessary. Enable all these languages
and test the functionality of the language selection menu in the footer.

- Non-enabled languages should appear as non-bold links.
- When a single language (with no sub-languages) is enabled it should
  appear as bold text.
- When one of multiple sub-languages is enabled that selection in the
  pop-up menu should appear as bold text. The menu label should show in
  bold the language code of the selected sub-language.

Tested with de-DE and de-CH, works as expected
Signed-off-by: Marc Véron <veron@veron.ch>

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Works as expected, passes all tests. No problems found.

Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>