Bug 12635 - Course reserves item list should use "Checked out" instead of "On Loan"
When viewing the list of items on a course reserve list, if the item is
checked out it is indicated by the phrase "On Loan." The default
templates should always use "checked out" instead of "on loan." Also
corrected: the capitalization of "Item Lost."

To test, Course Reserves must be enabled and you must have an existing
course reserves with items on it, one of which should be checked out.
View the list of items on that Course Reserve and confirm that the table
says "Checked out."

Table in staff client displays "Checked out" / "Item lost"
Signed-off-by: Marc Véron <veron@veron.ch>

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
String change, no problems found.

Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>