Bug 12728: [QA Follow-up] Add StatisticsFields to sysprefs.sql
This patch adds a dbrev.
If people installed Koha after the initial dbrev (3.09.00.015), an install
with sysprefs.sql did not add the pref. So they still could not have it.

Test plan:
If you have the pref, running an database update should not do anything.
If you do not have the pref (delete it via mysql), it should be added.

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>